pyhton/jinja2/requirements.txt
pyyaml == 5.3.1 
jinja2 == 3.0.3
